projects/verify_wikipedia/requirements.txt
hydra-core
jsonlines
omegaconf
torch
tqdm
transformers